pyproject.toml
[project]
name = "db-agent"
version = "0.1.0"
description = "A LangGraph-based agent application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]
